feat: Add Join page and update contact/give info
Introduces a new "Join" page for recruiting builders and updates contact and giving information to include postal addresses and recipient details for checks. Enhances the footer with a link to the new "Join" page and refactors UI elements for consistency.

Also updates the ticker text on the homepage to reflect the "Hiring Builders" call to action.